Move admin maintenance APIs into dedicated module
