+++ webapp/admin/__init__.py
@@ -9,6 +9,7 @@ from .api import register_api_routes
 from .auth import register_auth_routes
 from .coverage import calculate_coverage_percentages
 from .dashboard import register_dashboard_routes
+from .maintenance import register_maintenance_routes
 
 
 def register(app, logger):
@@ -18,6 +19,7 @@ def register(app, logger):
 
     register_audio_routes(app, logger, eas_config)
     register_api_routes(app, logger)
+    register_maintenance_routes(app, logger)
     register_auth_routes(app, logger)
     register_dashboard_routes(app, logger, eas_config)
